+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [bugfix] Print user-friendly message if user tries to open console on non
  application virtual server
+* [improvement] Add `add` alias for `shelly config create`
 
 ## 0.4.24 / 2014-02-25