+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# master
+## 0.4.25 / 2014-02-28
 
 * [bugfix] Print user-friendly message if user tries to open console on non
  application virtual server